Remove compiler warning in PlotView.cpp
Removes a compiler warning that was introduced by commit
6a3d9712bd9c2ed78dc4593fb7f4cf4ba15598d8.

+++ src/PlotView.cpp
@@ -280,7 +280,7 @@ void PlotView::smoothSpeed(double sigma)
 		c[x] /= c_sum;
 	}
 
-	size_t last = m_numPoints;
+	int last = m_numPoints;
 	double * data = m_curves[Speed].data();
 
 	// TODO: the first and the last two points are not smoothed
